Added YouTube parser functionality for retrieving video keywords
Adds functionality to the YouTube parser. The YouTube parser will now retrieve the keywords of a YouTube video which can be used in the template with %videoKeywords%.
My intention is to use these keywords as tags in Obsidian, therefore the output format is a string that looks like "#keyword1 #keyword2 #keyword3" etc.